requirements.txt
flask==1.1.1
line-bot-sdk==1.14.0
